Update messages in Message.cs and Program.cs
The `GetMessage` method in `Message.cs` was updated to return a new string: "and I am a message from Message.cs!" instead of "Hello from Message.cs!".

In `Program.cs`, the second `Console.WriteLine` message was updated to "This is a test project for the hackathon.".

Additionally, a call to `Message.GetMessage()` was added in `Program.cs`, though its result is not currently being used.